Direct method to submit the AjaxForm & VirtualAjaxForm
Now we can submit a form without any validations or beforeSubmit method
call.

$(element).ajaxForm().ajaxSubmit()
$(element).virtualAjaxForm().ajaxSubmit()

Useful when we want initial confirmation from user and then only want to
submit the form.

For Example, Do you really want to delete your account? In this case we
can use this method to directly submit the form.